% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjust_r_lower}
\alias{adjust_r_lower}
\title{adjust_r_lower}
\arguments{
\item{r1}{r parameter for the first group}

\item{p1}{p parameter for the first group}

\item{r2}{r parameter for the second group}

\item{p2}{p parameter for the second group}

\item{dn}{The observed number of difference between the two groups}
}
\description{
If one r value is too large for a pvalue to be calculated, the r value is adjusted until a p-value is calculable.
Generally only necessary in cases of extreme differential expression, hypergeometric function will not converge otherwise
This function operates for lower tailed tests.
}
