% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{calculate_k_subsample}
\alias{calculate_k_subsample}
\title{calculate_k_random}
\usage{
calculate_k_subsample(row, group)
}
\arguments{
\item{row}{counts for a single gene}

\item{group}{A factor that contains grouping information for the counts}
}
\value{
k for each gene
}
\description{
Calculate k based on the subsampling protocol
}
