% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{celltype.annotation}
\alias{celltype.annotation}
\title{celltype.annotation}
\usage{
celltype.annotation(sobj, celltype.min, output)
}
\arguments{
\item{sobj}{A seurat object containing normalized counts}

\item{celltype.min}{Filters out cells of celltypes that are present in too few cells.}

\item{output}{A Seruat object containing annotated counts}
}
\description{
performs celltype annotations with single R, if possible. Can be limited if dataset contains a small number of cells
}
