% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{clustering.w.singleR}
\alias{clustering.w.singleR}
\title{cluster.w.singleR}
\usage{
clustering.w.singleR(sobj, output)
}
\arguments{
\item{sobj}{A seurat object containing processed counts}

\item{output}{The type of output the user specified}
}
\value{
nothing, calls output
}
\description{
generates tsne coordinates for cells, then outputs the plot with cell types annotated
}
