% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{DEGage_complete}
\alias{DEGage_complete}
\title{DEGage_complete}
\usage{
DEGage_complete(
  input,
  dir.type = "mtx",
  min.nFeatureRNA = 200,
  max.nFeatureRNA = 8000,
  mt.percent = 0.2,
  cell.annotations = TRUE,
  celltype.min = 20,
  perm.preprocess = TRUE,
  gene.filter.threshold = 1,
  nperms = 2000,
  nsubsample = NA,
  perm.pval = 0.1,
  ncores = 4,
  maxiter = 100,
  mean.ratio = 1.4,
  subsampled.k = T,
  writing.dir = NULL
)
}
\arguments{
\item{input}{either a dataframe or matrix containing counts, or a path to an mtx directory}

\item{dir.type}{the type of directory that a path, if input, is directing to. Default is mtx.}

\item{min.nFeatureRNA}{The minimum number of genes that must be present in a cell}

\item{max.nFeatureRNA}{The maxiumum number of genes that can be present in a cell}

\item{mt.percent}{A value between 0 and 1 indicating the maximum proporition of mitochondrial genes allowed in a cell}

\item{cell.annotations}{A boolean indicating whether or not to perform celltype annotations.}

\item{celltype.min}{the minimum number of cells of a given cell type to count as valid.}

\item{perm.preprocess}{A logical indicating whether or not to perform the permutation prefiltering step. Setting it to TRUE increases runtimes.}

\item{gene.filter.threshold}{A value between 0-1 which represents the maximum proportion of zeros a gene can have before being filtered out.}

\item{nperms}{An integer greater than 0 that indicates how many permutations will be carried out during the permutation test}

\item{nsubsample}{The number of cells to subsample for each condition}

\item{perm.pval}{P value for the permutation test to filter out genes with}

\item{ncores}{the number of cores to allocated for parallel computing}

\item{maxiter}{The maxiumum number of iterations to perform while calculating the cdf}

\item{mean.ratio}{The minimum ratio between the count means for each group for a gene with an incalculable p-value to be considered differentially expressed}

\item{subsampled.k}{If true the subsampling procedure is used to estimate k. If false, the random assignment procedure is used.}

\item{writing.dir}{A directory to write the results of each comparison to. It is HIGHLY reccomended to provide a directory.}
}
\value{
If writing.dir is NULL, a list of dataframes containing the results for each comparison is returned. If a writing directory is provided, nothing is returned.
}
\description{
Performs cell type annotations and differential expression analysis for scRNA-seq count data
}
