% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage.R
\name{DEGage_mean}
\alias{DEGage_mean}
\title{DEGage_mean}
\usage{
DEGage_mean(r1, p1, r2, p2)
}
\arguments{
\item{r1}{r parameter for the first NB distribution}

\item{p1}{p parameter for the first NB distribution}

\item{r2}{r parameter for the second NB distrbution}

\item{p2}{p parameter for the second NB distribution}
}
\description{
Calculates mean of distrubtion
}
