% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{filter.and.normalize}
\alias{filter.and.normalize}
\title{filter.and.normalize}
\usage{
filter.and.normalize(sobj, min.nFeatureRNA, max.nFeatureRNA, mt.percent)
}
\arguments{
\item{sobj}{A Seurat object containing raw counts}

\item{min.nFeatureRNA}{The minimum number of genes that must be present in a cell}

\item{max.nFeatureRNA}{The maxiumum number of genes that can be present in a cell}

\item{mt.percent}{A value between 0 and 1 indicating the maximum proporition of mitochondrial genes allowed in a cell}
}
\value{
A seurat object containing raw counts
}
\description{
fiters and normalizes counts according to lightweight Seurat workflow
}
