% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_closest_k}
\alias{find_closest_k}
\title{find_closest_k}
\arguments{
\item{r1}{r parameter for the first group}
}
\description{
if k is too large for a pvalue to be calculable, it is adjusted until a pvalue can be calculated
}
