% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{fix_infs}
\alias{fix_infs}
\title{fix_infs}
\usage{
fix_infs(outputdf, mean.ratio)
}
\arguments{
\item{outputdf}{A dataframe containing regression parameters and pvalues}

\item{mean.ratio}{A positive value indicating the minimum mean ratio for an gene with an incalculable pvalue to be considered differentially expressed}
}
\description{
Fixes genes that could not be fit with DOTNB because of limitations with regression
or hypergeometric function calcualtions.
}
