% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage_Simulation.R
\name{gen_colnames}
\alias{gen_colnames}
\title{gen_colnames}
\usage{
gen_colnames(ncells, ngroup1, ncores)
}
\arguments{
\item{ncells}{the number of cells to be generated}

\item{ngroup1}{the number of cells in group one}

\item{ncores}{the number of cores to allocated for parallel computing}
}
\description{
A helper function that generates a name for each cell for simulation. Names indicate which condition/group each cell belongs to
}
