% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage_Simulation.R
\name{gen_rownames}
\alias{gen_rownames}
\title{gen_rownames}
\usage{
gen_rownames(ngenes, ndegs, ncores)
}
\arguments{
\item{ngenes}{the number of total genes}

\item{ndegs}{the number of desired differentially expressed genes}

\item{ncores}{the number of cores to allocated for parallel computing}
}
\description{
A helper function that generates row names for simulation.
}
