% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage_Simulation.R
\name{introduce_dropouts}
\alias{introduce_dropouts}
\title{gen_colnames}
\usage{
introduce_dropouts(finaldf, ncells, ndropouts)
}
\arguments{
\item{finaldf}{simulated count matrix}

\item{ncells}{the number of cells simulated}

\item{ndropouts}{the number of dropouts to be introduced}
}
\description{
A helper function that introduces dropouts at a predetermined volume
}
