% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{NB_model_fitting}
\alias{NB_model_fitting}
\title{NB_model_fitting}
\usage{
NB_model_fitting(counts, group)
}
\arguments{
\item{counts}{A sliced dataframe that contains the counts for an individual gene.}

\item{group}{A factor that contains grouping information for the counts}
}
\value{
A dataframe with a single row containing NB regression parameters for a gene
}
\description{
A helper function that performs genewise standard negative binomial regression
}
