% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{organize_output}
\alias{organize_output}
\title{organize_output}
\usage{
organize_output(outputdf)
}
\arguments{
\item{outputdf}{df with all information to be output to user}
}
\value{
An outputdf with the columns in a different order
}
\description{
formats output for user
}
