% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{output.format.w.celltypes}
\alias{output.format.w.celltypes}
\title{output.format.w.celltypes}
\usage{
output.format.w.celltypes(sobj, output)
}
\arguments{
\item{sobj}{Seurat objecct containing processed counts and cell type annotations}

\item{output}{format of output as specific by user}
}
\value{
filtered counts and cell type annotations as specified by user
}
\description{
formats output to the users preference
}
