% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{parse.input.preprocess}
\alias{parse.input.preprocess}
\title{parse.input}
\usage{
parse.input.preprocess(input, type)
}
\arguments{
\item{input}{User input of whatever format}

\item{type}{type of directory that a user input path leads to, if relevant.}
}
\description{
parses user input into a seurat object
}
