% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{permtest_facilitator}
\alias{permtest_facilitator}
\title{permtest_facilitator}
\usage{
permtest_facilitator(counts, group, nperms)
}
\arguments{
\item{counts}{A sliced dataframe that contains the counts for an individual gene.}

\item{group}{A factor that contains grouping information for the counts}

\item{nperms}{An integer which determines how many iterations are performed for the permutation test.}
}
\value{
a pvalue fora single gene
}
\description{
A helper function that calls upon the permutation test written in C++, runs a single gene at a time
}
