% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{preprocess.QC}
\alias{preprocess.QC}
\title{preprocess.QC}
\usage{
preprocess.QC(input, dir.type, output)
}
\arguments{
\item{input}{either a dataframe or matrix containing counts, or a path to an mtx directory}

\item{dir.type}{the type of directory that a path, if input, is directing to. Default is mtx.}

\item{output}{output format as specified by the user}
}
\description{
performs QC on user input
}
