% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multitest_functions.R
\name{print.labels}
\alias{print.labels}
\title{print.labels}
\usage{
\method{print}{labels}(labels)
}
\arguments{
\item{labels}{cell type annotations for every cell}
}
\value{
nothing
}
\description{
prints all labels identified after preprocessing
}
