% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGage_Simulation.R
\name{QC_simulation}
\alias{QC_simulation}
\title{QC_Simulation}
\usage{
QC_simulation(
  ngenes,
  ndegs,
  cellgroups,
  lfc = 1,
  prop.zeros = 0.3,
  seed = NULL,
  ncores = 4
)
}
\arguments{
\item{ngenes}{is the number of genes to be simulated total}

\item{ndegs}{is the number of genes to be differentially expressed}

\item{cellgroups}{is factor with integers corresponding to cell conditions}

\item{lfc}{is either an integer or vector with a length of ndegs which dictates log fold change value to adjust means for differentially expressed genes}

\item{prop.zeros}{The proportion of 0's to introduce to simulate dropouts. Must be a value between 0 and 1.}

\item{seed}{sets a seed for random generation. If user does not input one, a random seed is generated}

\item{ncores}{the number of cores to allocated for parallel computing}
}
\description{
Performs quality control on user input for DEGage_simulation
}
