% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multitest_functions.R
\name{run.multitest}
\alias{run.multitest}
\title{run.multitest}
\usage{
run.multitest(
  allcounts,
  group,
  perm.preprocess,
  gene.filter.threshold,
  nperms,
  nsubsample,
  perm.pval,
  ncores,
  writing.dir
)
}
\arguments{
\item{allcounts}{the entire matrix of counts for all cells}

\item{group}{A factor which assigns conditions to cells as in Deseq2.}

\item{perm.preprocess}{A logical indicating whether or not to perform the permutation prefiltering step. Setting it to TRUE increases runtimes.}

\item{gene.filter.threshold}{A value between 0-1 which represents the maximum proportion of zeros a gene can have before being filtered out.}

\item{nperms}{An integer greater than 0 that indicates how many permutations will be carried out during the permutation test}

\item{nsubsample}{The number of cells to subsample for each condition}

\item{perm.pval}{P value for the permutation test to filter out genes with}

\item{ncores}{the number of cores to allocated for parallel computing}

\item{writing.dir}{A directory to write the results of each comparison to. It is HIGHLY reccomended to provide a directory.}
}
\value{
Nothing if a directory is provided. If no directory is provided,
}
\description{
calls DEGage for every comparison necessary in the DEGage_complete workflow
}
