% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{run.permtest}
\alias{run.permtest}
\title{run.permtest}
\usage{
run.permtest(counts, group, perm.pval, nperms, cl)
}
\arguments{
\item{counts}{A sliced dataframe that contains the counts for an individual gene.}

\item{group}{A factor that contains grouping information for the counts}

\item{perm.pval}{Pvalues}

\item{nperms}{An integer which determines how many iterations are performed for the permutation test.}

\item{cl}{parallel computing cluster information}
}
\value{
A dataframe containing gene names with their corresponding pvalues
}
\description{
The first function called to initiate the permutation test, does not call c++ functions
}
