% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{subsampling}
\alias{subsampling}
\title{subsampling}
\usage{
subsampling(counts, group, nsubsample)
}
\arguments{
\item{counts}{A dataframe containing all read counts}

\item{group}{A factor containing column annotations}

\item{nsubsample}{Number of cells to subsample}
}
\value{
A list containing subsampled counts at index 1 and a revised group factor at index 2
}
\description{
Samples counts if the size of either group is greater than nsubsample
}
