% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{subsampling_helper}
\alias{subsampling_helper}
\title{subsampling_helper}
\usage{
subsampling_helper(counts, group, nsubsample, n)
}
\arguments{
\item{counts}{A dataframe containing all read counts}

\item{group}{A factor containing column annotations}

\item{nsubsample}{Number of cells to subsample}

\item{n}{group number. Either 1 or 2.}
}
\value{
A dataframe containing containing sliced counts for one group
}
\description{
slices counts by randomly sampling columns
}
