% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{core.config}
\alias{core.config}
\title{core.config}
\usage{
core.config(ncores, nperms, counts, group)
}
\arguments{
\item{ncores}{Number of cores to use}
}
\value{
A cluster object to use for future parallel computing
}
\description{
configures cores for parallel computing
}
