% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{DEGage_preprocess}
\alias{DEGage_preprocess}
\title{DEGage_preprocess}
\usage{
DEGage_preprocess(
  input,
  dir.type = "mtx",
  min.nFeatureRNA = 200,
  max.nFeatureRNA = 8000,
  mt.percent = 0.2,
  cell.annotations = TRUE,
  celltype.min = 20,
  output = "Seurat"
)
}
\arguments{
\item{input}{either a dataframe or matrix containing counts, or a path to an mtx directory}

\item{dir.type}{the type of directory that a path, if input, is directing to. Can be 'mtx'.}

\item{min.nFeatureRNA}{The minimum number of genes that must be present in a cell}

\item{max.nFeatureRNA}{The maxiumum number of genes that can be present in a cell}

\item{mt.percent}{A value between 0 and 1 indicating the maximum proporition of mitochondrial genes allowed in a cell}

\item{cell.annotations}{A boolean indicating whether or not to perform celltype annotations.}

\item{celltype.min}{the minimum number of cells of a given cell type to count as valid.}

\item{output}{Format of the output. Can be "Seurat" or "df"}
}
\value{
A data structure of designated user input that contains count information
}
\description{
A simple preprocessing workflow to preprocess and generate SingleR cell type
annotations for scRNA-seq data.
}
