% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_Functions.R
\name{filter.and.normalize}
\alias{filter.and.normalize}
\title{filter.and.normalize}
\usage{
filter.and.normalize(sobj, min.nFeatureRNA, max.nFeatureRNA, mt.percent)
}
\arguments{
\item{sobj}{A Seurat object containing raw counts}
}
\value{
A seurat object containing raw counts
}
\description{
fiters and normalizes counts according to lightweight Seurat workflow
}
