% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hypergeo}
\alias{hypergeo}
\title{hypergeo}
\arguments{
\item{a}{Upper bound 1}

\item{b}{Upper bound 2}

\item{c}{lower bound}

\item{x}{primary complex argument}
}
\description{
calculates an approximation for the hypergeometric function for use in DEGage_pdf. Equivalent
to a 2F1 hypergeometric function calcuation with series expansion.
}
