% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{parse.input}
\alias{parse.input}
\title{parse.input}
\usage{
parse.input(counts)
}
\arguments{
\item{counts}{user input for 'counts'}
}
\value{
dataframe containing counts
}
\description{
parses input of the user based on the data type of the input
}
