% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Helper_functions.R
\name{pdf_facilitator}
\alias{pdf_facilitator}
\title{pdf_facilitator}
\usage{
pdf_facilitator(df, ncores, cl)
}
\arguments{
\item{df}{The data frame containing regression parameters for every gene}

\item{ncores}{number of cores to used}

\item{cl}{Parallel computing cluster info}
}
\description{
A helper function that calls upon DEGage_pdf
}
